+++ PA2/include/cellSpreader.h
@@ -5,8 +5,8 @@
 
 using namespace std;
 
-vector < vector <vector <int>>> initBinMatrix(vector <double> &XVector, vector <double> &YVector, int circuitWidth, int circuitHeight, int k);
-void XSpreadCells(vector <double> &XSpreadedVector, vector <double> &XVector, vector <double> &YVector, int circuitWidth, int circuitHeight, int k);
-void YSpreadCells(vector <double> &YSpreadedVector, vector <double> &XVector, vector <double> &YVector, int circuitWidth, int circuitHeight, int k);
+vector < vector <vector <int>>> initBinMatrix(vector <float> &XVector, vector <float> &YVector, int circuitWidth, int circuitHeight, int k);
+void XSpreadCells(vector <float> &XSpreadedVector, vector <float> &XVector, vector <float> &YVector, int circuitWidth, int circuitHeight, int k);
+void YSpreadCells(vector <float> &YSpreadedVector, vector <float> &XVector, vector <float> &YVector, int circuitWidth, int circuitHeight, int k);
 
 #endif
